+++ index.d.ts
@@ -21,6 +21,7 @@ declare namespace nkruntime {
         env: {[key: string]: string},
         executionMode: string,
         node: string,
+        headers: {[key: string]: string[]},
         queryParams: {[key: string]: string[]},
         userId: string,
         username: string,
@@ -32,7 +33,8 @@ declare namespace nkruntime {
         matchId: string,
         matchNode: string,
         matchLabel: string,
-        matchTickRate: number
+        matchTickRate: number,
+        lang: string,
     }
 
     type ReadPermissionValues = 0 | 1 | 2;
@@ -150,9 +152,10 @@ declare namespace nkruntime {
          * @param ctx - The context for the execution.
          * @param logger - The server logger.
          * @param nk - The Nakama server APIs.
-         * @param envelope - The Envelope message received by the function.
+         * @param output - The response envelope, if any.
+         * @param input - The Envelope message received by the function.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, envelope: T): void;
+        (ctx: Context, logger: Logger, nk: Nakama, output: T | null, input: T): void;
     }
 
     /**
@@ -238,7 +241,7 @@ declare namespace nkruntime {
          * @param reliable - Opt. Broadcast the message with delivery guarantees or not. Defaults to true.
          * @throws {TypeError, GoError}
          */
-        broadcastMessage(opcode: number, data?: string | null, presences?: Presence[] | null, sender?: Presence | null, reliable?: boolean): void;
+        broadcastMessage(opcode: number, data?: Uint8Array | string | null, presences?: Presence[] | null, sender?: Presence | null, reliable?: boolean): void;
 
         /**
          * Defer message broadcast to match presences.
@@ -250,7 +253,7 @@ declare namespace nkruntime {
          * @param reliable - Opt. Broadcast the message with delivery guarantees or not. Defaults to true.
          * @throws {TypeError, GoError}
          */
-        broadcastMessageDeferred(opcode: number, data?: string | null, presences?: Presence[] | null, sender?: Presence, reliable?: boolean): void;
+        broadcastMessageDeferred(opcode: number, data?: Uint8Array | string | null, presences?: Presence[] | null, sender?: Presence, reliable?: boolean): void;
 
         /**
          * Kick presences from match.
@@ -279,7 +282,7 @@ declare namespace nkruntime {
         persistence: boolean;
         status: string;
         opCode: number;
-        data: string;
+        data: Uint8Array;
         reliable: boolean;
         receiveTime: number;
     }
@@ -794,19 +797,20 @@ declare namespace nkruntime {
     /**
      * Match handler definitions
      */
-    export interface MatchHandler {
-        matchInit: MatchInitFunction;
-        matchJoinAttempt: MatchJoinAttemptFunction;
-        matchJoin: MatchJoinFunction;
-        matchLeave: MatchLeaveFunction;
-        matchLoop: MatchLoopFunction;
-        matchTerminate: MatchTerminateFunction;
+    export interface MatchHandler<State = MatchState> {
+        matchInit: MatchInitFunction<State>;
+        matchJoinAttempt: MatchJoinAttemptFunction<State>;
+        matchJoin: MatchJoinFunction<State>;
+        matchLeave: MatchLeaveFunction<State>;
+        matchLoop: MatchLoopFunction<State>;
+        matchTerminate: MatchTerminateFunction<State>;
+        matchSignal: MatchSignalFunction<State>;
     }
 
     /**
      * Match initialization function definition.
      */
-    export interface MatchInitFunction {
+    export interface MatchInitFunction<State = MatchState> {
         /**
          * Match initialization function definition.
          * @param ctx - The context for the execution.
@@ -815,13 +819,13 @@ declare namespace nkruntime {
          * @param params - Match create http request parameters.
          * @returns An object with the match state, tick rate and labels.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, params: {[key: string]: string}): {state: MatchState, tickRate: number, label: string};
+        (ctx: Context, logger: Logger, nk: Nakama, params: {[key: string]: string}): {state: State, tickRate: number, label: string};
     }
 
     /**
      * Match join attempt function definition.
      */
-    export interface MatchJoinAttemptFunction {
+    export interface MatchJoinAttemptFunction<State = MatchState> {
         /**
          * User match join attempt function definition.
          * @param ctx - The context for the execution.
@@ -834,13 +838,13 @@ declare namespace nkruntime {
          * @param metadata - Metadata object.
          * @returns object with state, acceptUser and optional rejection message if acceptUser is false.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: MatchState, presence: Presence, metadata: {[key: string]: any}): {state: MatchState, accept: boolean, rejectMessage?: string} | null;
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, presence: Presence, metadata: {[key: string]: any}): {state: State, accept: boolean, rejectMessage?: string} | null;
     }
 
     /**
      * Match join function definition.
      */
-    export interface MatchJoinFunction {
+    export interface MatchJoinFunction<State = MatchState> {
         /**
          * User match join function definition.
          * @param ctx - The context for the execution.
@@ -852,13 +856,13 @@ declare namespace nkruntime {
          * @param presences - List of presences.
          * @returns object with the new state of the match.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: MatchState, presences: Presence[]): {state: MatchState} | null;
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, presences: Presence[]): {state: State} | null;
     }
 
     /**
      * Match leave function definition.
      */
-    export interface MatchLeaveFunction {
+    export interface MatchLeaveFunction<State = MatchState> {
         /**
          * User match leave function definition.
          * @param ctx - The context for the execution.
@@ -870,13 +874,13 @@ declare namespace nkruntime {
          * @param presences - List of presences.
          * @returns object with the new state of the match.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: MatchState, presences: Presence[]): {state: MatchState} | null;
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, presences: Presence[]): {state: State} | null;
     }
 
     /**
      * Match loop function definition.
      */
-    export interface MatchLoopFunction {
+    export interface MatchLoopFunction<State = MatchState> {
         /**
          * User match leave function definition.
          * @param ctx - The context for the execution.
@@ -887,13 +891,13 @@ declare namespace nkruntime {
          * @param state - Current match state.
          * @param messages - Received messages in the buffer.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: MatchState, messages: MatchMessage[]): {state: MatchState} | null;
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, messages: MatchMessage[]): {state: State} | null;
     }
 
     /**
      * Match terminate function definition.
      */
-    export interface MatchTerminateFunction {
+    export interface MatchTerminateFunction<State = MatchState> {
         /**
          * User match leave function definition.
          * @param ctx - The context for the execution.
@@ -904,7 +908,25 @@ declare namespace nkruntime {
          * @param state - Current match state.
          * @param graceSeconds - Number of seconds to gracefully terminate the match. If this time elapses before the function returns the match will be forcefully terminated.
          */
-        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: MatchState, graceSeconds: number): {state: MatchState} | null;
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, graceSeconds: number): {state: State} | null;
+    }
+
+    /**
+     * Match signal function definition.
+     */
+    export interface MatchSignalFunction<State = MatchState> {
+        /**
+         * User match leave function definition.
+         * @param ctx - The context for the execution.
+         * @param logger - The server logger.
+         * @param nk - The Nakama server APIs.
+         * @param dispatcher - Message dispatcher APIs.
+         * @param tick - Current match loop tick.
+         * @param state - Current match state.
+         * @param data - Arbitrary data the signal caller is sending to the match signal handler.
+         * @returns object with state and optional response data string to the signal caller.
+         */
+        (ctx: Context, logger: Logger, nk: Nakama, dispatcher: MatchDispatcher, tick: number, state: State, data: string): {state: State, data?: string} | null;
     }
 
     /**
@@ -1026,9 +1048,9 @@ declare namespace nkruntime {
         registerBeforeAuthenticateEmail(fn: BeforeHookFunction<AuthenticateEmailRequest>): void;
 
         /**
-         * Register after Hook for RPC uthenticateEmail function.
+         * Register after Hook for RPC AuthenticateEmail function.
          *
-         * @param fn - The function to execute after uthenticateEmail.
+         * @param fn - The function to execute after AuthenticateEmail.
          * @throws {TypeError}
          */
         registerAfterAuthenticateEmail(fn: AfterHookFunction<Session, AuthenticateEmailRequest>): void;
@@ -2010,6 +2032,24 @@ declare namespace nkruntime {
          registerAfterValidatePurchaseApple(fn: AfterHookFunction<ValidatePurchaseResponse, ValidatePurchaseAppleRequest>): void;
 
         /**
+         * Register before Hook for RPC ValidateSubscriptionApple function.
+         *
+         * @param fn - The function to execute before ValidatePurchaseApple.
+         * @throws {TypeError}
+         */
+        registerBeforeValidateSubscriptionApple(fn: BeforeHookFunction<ValidateSubscriptionAppleRequest>): void;
+
+        /**
+         * Register after Hook for RPC ValidatePurchaseApple function.
+         *
+         * @param fn - The function to execute after ValidatePurchaseApple.
+         * @throws {TypeError}
+         */
+        registerAfterValidateSubscriptionApple(fn: AfterHookFunction<ValidateSubscriptionResponse, ValidateSubscriptionAppleRequest>): void;
+
+
+
+        /**
          * Register before Hook for RPC ValidatePurchaseGoogle function.
          *
          * @param fn - The function to execute before ValidatePurchaseGoogle.
@@ -2024,6 +2064,22 @@ declare namespace nkruntime {
          * @throws {TypeError}
          */
          registerAfterValidatePurchaseGoogle(fn: AfterHookFunction<ValidatePurchaseResponse, ValidatePurchaseGoogleRequest>): void;
+
+        /**
+         * Register before Hook for RPC ValidatePurchaseGoogle function.
+         *
+         * @param fn - The function to execute before ValidatePurchaseGoogle.
+         * @throws {TypeError}
+         */
+        registerBeforeValidatePurchaseGoogle(fn: BeforeHookFunction<ValidateSubscriptionGoogleRequest>): void;
+
+        /**
+         * Register after Hook for RPC ValidatePurchaseGoogle function.
+         *
+         * @param fn - The function to execute after ValidatePurchaseGoogle.
+         * @throws {TypeError}
+         */
+        registerAfterValidatePurchaseGoogle(fn: AfterHookFunction<ValidateSubscriptionResponse, ValidateSubscriptionGoogleRequest>): void;
 
         /**
          * Register before Hook for RPC ValidatePurchaseHuawei function.
@@ -2063,7 +2119,7 @@ declare namespace nkruntime {
          * @param name - Identifier of the match handler.
          * @param functions - Object containing the match handler functions.
          */
-        registerMatch(name: string, functions: MatchHandler): void;
+        registerMatch<State = MatchState>(name: string, functions: MatchHandler<State>): void;
 
         /**
          * Register matchmaker matched handler.
@@ -2306,6 +2362,7 @@ declare namespace nkruntime {
         matchId: string;
         authoritative: boolean;
         size: number;
+        label: string;
     }
 
     /**
@@ -2395,7 +2452,7 @@ declare namespace nkruntime {
     export interface StorageWriteRequest {
         key: string;
         collection: string;
-        userId: string;
+        userId: string | undefined;
         value: {[key: string]: any};
         version?: string;
         permissionRead?: ReadPermissionValues;
@@ -2501,6 +2558,7 @@ declare namespace nkruntime {
         maxCount: number;
         createTime: number;
         updateTime: number;
+        metadata: {[key: string]: any};
     }
 
     export interface UserGroupList {
@@ -2641,6 +2699,7 @@ declare namespace nkruntime {
         query: string
         stringProperties: {[key: string]: string}
         numericProperties: {[key: string]: number}
+        countMultiple: number
     }
 
     export interface EnvelopeMatchmakerAdd {
@@ -2703,8 +2762,16 @@ declare namespace nkruntime {
         receipt: string
     }
 
+    export interface ValidateSubscriptionAppleRequest {
+        receipt: string
+    }
+
     export interface ValidatePurchaseGoogleRequest {
         purchase: string
+    }
+
+    export interface ValidateSubscriptionGoogleRequest {
+        receipt: string
     }
 
     export interface ValidatePurchaseHuaweiRequest {
@@ -2716,9 +2783,18 @@ declare namespace nkruntime {
         validatedPurchases?: ValidatedPurchase[]
     }
 
+    export interface ValidateSubscriptionResponse {
+        validatedSubscription: ValidatedSubscription
+    }
+
     export interface ValidatedPurchaseOwner {
-        validatedPurchase: ValidatedPurchase,
-        userId: string,
+        validatedPurchase: ValidatedPurchase
+        userId: string
+    }
+
+    export interface ValidatedSubscriptionOwner {
+      validatedSubscription: ValidatedSubscription
+      userId: string
     }
 
     export type ValidatedPurchaseStore = "APPLE_APP_STORE" | "GOOGLE_PLAY_STORE" | "HUAWEI_APP_GALLERY"
@@ -2732,14 +2808,33 @@ declare namespace nkruntime {
         purchaseTime: string
         createTime: string
         updateTime: string
-        providerPayload: string
+        providerResponse: string
         environment: ValidatedPurchaseEnvironment
         seenBefore: boolean
+    }
+
+    export interface ValidatedSubscription {
+        productId: string
+        originalTransactionId: string
+        store: ValidatedPurchaseStore
+        purchaseTime: string
+        createTime: string
+        updateTime: string
+        environment: ValidatedPurchaseEnvironment
+        expiryTime: string
+        active: boolean
     }
 
     export interface ValidatedPurchaseList {
         validatedPurchases?: ValidatedPurchase[]
         cursor?: string
+        prevCursor?: string
+    }
+
+    export interface ValidatedSubscriptionList {
+        validatedSubscription?: ValidatedSubscription
+        cursor?: string
+        prevCursor?: string
     }
 
     export interface ChannelMessageSendAck {
@@ -2771,6 +2866,22 @@ declare namespace nkruntime {
      */
     export interface Nakama {
         /**
+         * Convert binary data to string.
+         *
+         * @param data - Data to convert to string.
+         * @throws {TypeError}
+         */
+         binaryToString(data: Uint8Array): string;
+
+        /**
+         * Convert a string to binary data.
+         *
+         * @param str - String to convert to binary data.
+         * @throws {TypeError}
+         */
+         stringToBinary(str: string): Uint8Array;
+
+        /**
          * Emit an event to be processed.
          *
          * @param eventName - A string with the event name.
@@ -2780,6 +2891,37 @@ declare namespace nkruntime {
          * @throws {TypeError}
          */
         event(eventName: string, properties: {[key: string]: string}, timestamp?: number, external?: boolean): void;
+
+        /**
+         * Add a custom metrics counter.
+         *
+         * @param name - The name of the custom metrics counter.
+         * @param tags - The metrics tags associated with this counter.
+         * @param delta - An integer value to update this metric with.
+         * @throws {TypeError}
+         */
+        metricsCounterAdd(name: string, tags: {[key: string]: string}, delta: number): void;
+
+        /**
+         * Add a custom metrics gauge.
+         *
+         * @param name - The name of the custom metrics gauge.
+         * @param tags - The metrics tags associated with this gauge.
+         * @param value - A value to update this metric with.
+         * @throws {TypeError}
+         */
+        metricsGaugeSet(name: string, tags: {[key: string]: string}, value: number): void;
+
+
+        /**
+         * Add a custom metrics timer.
+         *
+         * @param name - The name of the custom metrics timer.
+         * @param tags - The metrics tags associated with this timer.
+         * @param value - An integer value to update this metric with (in nanoseconds).
+         * @throws {TypeError}
+         */
+        metricsTimerRecord(name: string, tags: {[key: string]: string}, value: number): void;
 
         /**
          * Generate a new UUID v4.
@@ -3510,7 +3652,17 @@ declare namespace nkruntime {
          * @param sessionID - Opt. Presence disconnect reason.
          * @throws {TypeError, GoError}
          */
-        sessionDisconnect(sessionID: string, reason?: PresenceReason): void;
+         sessionDisconnect(sessionID: string, reason?: PresenceReason): void;
+
+        /**
+         * Log out a user from their current session.
+         *
+         * @param userId - The ID of the user to be logged out.
+         * @param token - Opt. The current session authentication token.
+         * @param refreshToken - Opt. The current session refresh token.
+         * @throws {TypeError, GoError}
+         */
+         sessionLogout(userId: string, token?: string, refreshToken?:string): void;
 
         /**
          * Create a new match.
@@ -3525,7 +3677,7 @@ declare namespace nkruntime {
         /**
          * Get a running match info.
          *
-         * @param matchID - Match ID.
+         * @param id - Match ID.
          * @returns match data.
          * @throws {TypeError, GoError}
          */
@@ -3544,6 +3696,17 @@ declare namespace nkruntime {
          * @throws {TypeError, GoError}
          */
         matchList(limit: number, authoritative?: boolean | null, label?: string | null, minSize?: number | null, maxSize?: number | null, query?: string | null): Match[]
+
+        /**
+         * Signal a match and receive a response.
+         *
+         * @param id - Match ID.
+         * @param data - Arbitrary data to pass to the match signal handler.
+         * @returns response data from the signal handler, if any.
+         * @throws {TypeError, GoError}
+         */
+        matchSignal(id: string, data: string): string
+
 
         /**
          * Send a notification.
@@ -3565,6 +3728,17 @@ declare namespace nkruntime {
          * @throws {TypeError, GoError}
          */
         notificationsSend(notifications: NotificationRequest[]): void;
+
+        /**
+         * Send an in-app notification to all users.
+         *
+         * @param subject - Subject of the notification.
+         * @param content - Key value object to send as the notification content.
+         * @param code - Custom code for the notification. Must be a positive integer.
+         * @param persistent - Opt. A non-persistent message will only be received by a client which is currently connected to the server. Defaults to false.
+         * @throws {TypeError, GoError}
+         */
+        notificationSendAll(subject: string, content: {[key: string]: any}, code: number, persistent?: boolean): void;
 
         /**
          * Update user wallet.
@@ -3746,6 +3920,18 @@ declare namespace nkruntime {
         leaderboardsGetId(leaderboardIds: string[]): Leaderboard[];
 
         /**
+         * Fetch the list of leaderboard records around the owner.
+         *
+         * @param leaderboardId - The unique identifier for the leaderboard.
+         * @param ownerId - The owner of the score to list records around. Mandatory field.
+         * @param limit - Return only the required number of leaderboard records denoted by this limit value.
+         * @param overrideExpiry - Records with expiry in the past are not returned unless within this defined limit. Must be equal or greater than 0.
+         * @returns The leaderboard records according to ID.
+         * @throws {TypeError, GoError}
+         */
+        leaderboardRecordsHaystack(leaderboardId: string, ownerId: string, limit: number, overrideExpiry: number): LeaderboardRecordList[];
+
+        /**
          * Create a new tournament.
          *
          * @param tournamentID - Tournament id.
@@ -3770,7 +3956,7 @@ declare namespace nkruntime {
             authoritative: boolean,
             sortOrder: SortOrder,
             operator: Operator,
-            duration: number,
+            duration?: number,
             resetSchedule?: string | null,
             metadata?: {[key: string]: any} | null,
             title?: string | null,
@@ -3981,6 +4167,39 @@ declare namespace nkruntime {
         friendsList(userId: string, limit?: number, state?: number, cursor?: string): FriendList;
 
         /**
+         * Add friends to a user.
+         *
+         * @param userId - User ID.
+         * @param username - Username.
+         * @param ids - The IDs of the users you want to add as friends.
+         * @param usernames - The usernames of the users you want to add as friends.
+         * @throws {TypeError, GoError}
+         */
+        friendsAdd(userId: string, username: string, ids: string[], usernames: string[]): FriendList;
+
+        /**
+         * Delete friends from a user.
+         *
+         * @param userId - User ID.
+         * @param username - Username.
+         * @param ids - The IDs of the users you want to delete as friends.
+         * @param usernames - The usernames of the users you want to delete as friends.
+         * @throws {TypeError, GoError}
+         */
+        friendsDelete(userId: string, username: string, ids: string[], usernames: string[]): FriendList;
+
+        /**
+         * Block friends for a user.
+         *
+         * @param userId - User ID.
+         * @param username - Username.
+         * @param ids - The IDs of the users you want to block as friends.
+         * @param usernames - The usernames of the users you want to block as friends.
+         * @throws {TypeError, GoError}
+         */
+        friendsBlock(userId: string, username: string, ids: string[], usernames: string[]): FriendList;
+
+        /**
          * Join a user to a group.
          *
          * @param groupID - Group ID.
@@ -4009,6 +4228,16 @@ declare namespace nkruntime {
          * @throws {TypeError, GoError}
          */
         groupUsersAdd(groupID: string, userIds: string[], callerID?: string): void;
+
+        /**
+         * Ban multiple users from a group.
+         *
+         * @param groupID - Group ID.
+         * @param userIds - Array of userIds to ban from the group.
+         * @param callerID - Opt. User ID mandating the operation to check for sufficient priviledges. Defaults to admin user if empty.
+         * @throws {TypeError, GoError}
+         */
+         groupUsersBan(groupID: string, userIds: string[], callerID?: string): void;
 
         /**
          * Promote users in a group.
@@ -4052,31 +4281,36 @@ declare namespace nkruntime {
          *
          * @param userID - User ID.
          * @param receipt - Apple receipt to validate.
+         * @param persist - Opt. Whether to persist the receipt validation. Defaults to true.
          * @param passwordOverride - Opt. Override the configured Apple Store Validation Password.
          * @returns The result of the validated and stored purchases from the receipt.
          * @throws {TypeError, GoError}
          */
-        purchaseValidateApple(userID: string, receipt: string, passwordOverride?: string): ValidatePurchaseResponse
+        purchaseValidateApple(userID: string, receipt: string, persist?: boolean, passwordOverride?: string): ValidatePurchaseResponse
 
         /**
          * Validate a Google purchase payload.
          *
          * @param userID - User ID.
          * @param purchase - Google purchase payload to validate.
+         * @param persist - Opt. Whether to persist the receipt validation. Defaults to true.
+         * @param clientEmailOverride - Opt. Override the configured Google Service Account client email.
+         * @param privateKeyOverride - Opt. Override the configured Google Service Account private key.
          * @returns The result of the validated and stored purchases from the receipt.
          * @throws {TypeError, GoError}
          */
-         purchaseValidateGoogle(userID: string, purchase: string): ValidatePurchaseResponse
+         purchaseValidateGoogle(userID: string, purchase: string, persist?: boolean, clientEmailOverride?: string, privateKeyOverride?: string): ValidatePurchaseResponse
 
         /**
          * Validate a Huawei purchase payload.
          *
          * @param userID - User ID.
          * @param receipt - Apple receipt to validate.
+         * @param persist - Opt. Whether to persist the receipt validation. Defaults to true.
          * @returns The result of the validated and stored purchases from the receipt.
          * @throws {TypeError, GoError}
          */
-        purchaseValidateHuawei(userID: string, receipt: string, signature: string): ValidatePurchaseResponse
+        purchaseValidateHuawei(userID: string, receipt: string, signature: string, persist?: boolean): ValidatePurchaseResponse
 
         /**
          * Get a validated purchase data by transaction ID.
@@ -4086,6 +4320,52 @@ declare namespace nkruntime {
          * @throws {TypeError, GoError}
          */
         purchaseGetByTransactionId(transactionID: string): ValidatedPurchaseOwner
+
+        /**
+         * List validated and stored purchases.
+         *
+         * @param userID - Opt. User ID.
+         * @param limit - Opt. Limit of results per page. Must be a value between 1 and 100.
+         * @param cursor - Opt. A cursor used to fetch the next page when applicable.
+         * @returns A page of validated and stored purchases.
+         * @throws {TypeError, GoError}
+         */
+        purchasesList(userID?: string, limit?: number, cursor?: string): ValidatedPurchaseList
+
+        /**
+         * Validate an Apple receipt containing a subscription.
+         *
+         * @param userID - User ID.
+         * @param receipt - Apple subscription receipt to validate.
+         * @param persist - Opt. Whether to persist the subscription validation. Defaults to true.
+         * @param passwordOverride - Opt. Override the configured Apple Store Validation Password.
+         * @returns The result of the validated and stored purchases from the receipt.
+         * @throws {TypeError, GoError}
+         */
+        subscriptionValidateApple(userID: string, receipt: string, persist?: boolean, passwordOverride?: string): ValidateSubscriptionResponse
+
+        /**
+         * Validate a Google receipt containing a subscription.
+         *
+         * @param userID - User ID.
+         * @param subscription - Google subscription payload to validate.
+         * @param persist - Opt. Whether to persist the subscription receipt validation. Defaults to true.
+         * @param clientEmailOverride - Opt. Override the configured Google Service Account client email.
+         * @param privateKeyOverride - Opt. Override the configured Google Service Account private key.
+         * @returns The result of the validated and stored purchases from the receipt.
+         * @throws {TypeError, GoError}
+         */
+        subscriptionValidateGoogle(userID: string, subscription: string, persist?: boolean, clientEmailOverride?: string, privateKeyOverride?: string): ValidateSubscriptionResponse
+
+        /**
+         * Get a validated subscription data by product ID.
+         *
+         * @param userID - User ID.
+         * @param productID - Product ID. For Google this is the subscriptionToken value of the purchase data.
+         * @returns The data of the validated and stored purchase.
+         * @throws {TypeError, GoError}
+         */
+        subscriptionGetByProductId(userID: string, productID: string): ValidatedPurchaseOwner
 
         /**
          * List validated and stored purchases.
@@ -4126,14 +4406,37 @@ declare namespace nkruntime {
          channelMessageUpdate(channelId: string, messageId: string, content?: {[key: string]: any}, senderId?: string, senderUsername?: string, persist?: boolean): ChannelMessageSendAck
 
         /**
+         * List channel messages.
+         *
+         * @param channelId - Channel ID.
+         * @param limit - The number of messages to return per page.
+         * @param forward - Whether to list messages from oldest to newest, or newest to oldest.
+         * @param cursor - Opt. Pagination cursor.
+         * @returns List of channel messages.
+         * @throws {TypeError, GoError}
+         */
+         channelMessagesList(channelId: string, limit?: number, forward?: boolean, cursor?: string): ChannelMessageList
+
+        /**
          * Send channel message.
          *
+         * @param sender - The user ID of the sender.
          * @param target - The user ID to DM with, group ID to chat with, or room channel name to join.
          * @param type - Channel type.
          * @returns The channelId.
          * @throws {TypeError, GoError}
          */
-        channelIdBuild(target: string, chanType: ChanType): string
+        channelIdBuild(sender: string, target: string, chanType: ChanType): string
+
+        /**
+         * Parses a CRON expression and a timestamp in UTC seconds, and returns the next matching timestamp in UTC seconds.
+         *
+         * @param cron - The cron expression.
+         * @param timestamp - UTC unix seconds timestamp.
+         * @returns The next cron matching timestamp in UTC seconds.
+         * @throws {TypeError, GoError}
+         */
+         cronNext(cron: string, timestamp: number): number
     }
 
     /**